use delta time in movement calculations

+++ core/src/uniform/sprite.rs
@@ -3,9 +3,9 @@ use bytemuck::{Pod, Zeroable};
 #[derive(Default, Copy, Clone, Pod, Zeroable)]
 #[repr(C)]
 pub struct Sprite {
-    pub x: i32,
-    pub y: i32,
-    pub width: i32,
-    pub height: i32,
+    pub x: f32,
+    pub y: f32,
+    pub width: f32,
+    pub height: f32,
     pub texture_clip: [f32; 4],
 }
